# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel HRM Documentation",
        "description": "L5 Swagger OpenApi description for hrm project",
        "contact": {
            "email": "drrifatalashwad0@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "enrollments": {
            "post": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Enroll authenticated user in a course",
                "operationId": "46a10ea4003d6986b47c28dc399c767c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 101
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enrollment successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "course_id": {
                                            "type": "integer",
                                            "example": 101
                                        },
                                        "enrolled_at": {
                                            "type": "string",
                                            "example": "2025-09-16 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "users/{id}/enrollments": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Get all enrollments of a specific user",
                "operationId": "19ca0bf3f75742f6b52f2404c0418319",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user enrollments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "enrollment_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "course": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 101
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Laravel Basics"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Learn Laravel framework"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": 49.99
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "progress": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "completed"
                                            },
                                            "enrolled_at": {
                                                "type": "string",
                                                "example": "2025-09-16 12:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "courses/{course_id}/lessons": {
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Create a new lesson for a course (Admin only)",
                "operationId": "cb72058b07ecfa0d8b985ec390714d33",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content_type",
                                    "content_url"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Introduction to Laravel"
                                    },
                                    "content_type": {
                                        "type": "string",
                                        "example": "video"
                                    },
                                    "content_url": {
                                        "type": "string",
                                        "example": "https://example.com/video.mp4"
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lesson created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "lessons/{id}": {
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Update an existing lesson (Admin only)",
                "operationId": "aa67ac7b5855fbde7b6446184970ebc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Lesson Title"
                                    },
                                    "content_type": {
                                        "type": "string",
                                        "example": "video"
                                    },
                                    "content_url": {
                                        "type": "string",
                                        "example": "https://example.com/updated-video.mp4"
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson updated successfully"
                    },
                    "404": {
                        "description": "Lesson not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}